==> Asurion-customer-propensity/10-process-data.ipynb <==
# 10-process-data
> Importing, cleaning, testing, and saving data

This series of notebooks reflects operations to scrape, prepare, validate, and save the data.

#### Helpful packages and preliminaries

#data access and processing
import pandas as pd
import numpy as np

#Testing assertions
import great_expectations as ge

data_in_filename = ''
cleaned_data_filename = ''
