==> Asurion-customer-propensity/11-join-function.ipynb <==
import pandas as pd

%cd ..

march_encoded = pd.read_csv("data/tabular_march_encoded.csv")
march_collated = pd.read_csv("data/transcript_march_collated_encoded.csv")

def join_func(df1, df2, col):
    df_j = df1.merge(df2, on = col, how = "inner")
    return df_j

march_total = join_func(march_encoded, march_collated, "ID")

encoded_retained = len(march_total)/len(march_encoded)
print("%.2f percent of the tabular march encoded data were retained."%(encoded_retained*100))

collated_retained = len(march_total)/len(march_collated)
print("%.2f percent of the transcript march collated data were retained."%(collated_retained*100))
