==> Asurion-customer-propensity/12-join-function2.ipynb <==
import pandas as pd

%cd ..

total_transcripts = pd.read_csv("fe_data/total_collated_transcripts.csv")

total_tabular = pd.read_csv("fe_data/total_encoded_tabular.csv")

total_combined = total_tabular.merge(total_transcripts, how = "inner", on = ["id","month"])

total_combined.info()

total_combined = total_combined.drop(["accepted_flg_y"], axis = 1)
total_combined = total_combined.rename({"accepted_flg_x":"accepted_flg"})

total_combined.info()

total_combined.to_csv("fe_data/total_merged_asurion.csv", index = False)
