==> Asurion-customer-propensity/20-data-exploration.ipynb <==
# 20-data-exploration
> Exploratory data analysis

The purpose of this notebook series is to perform any desired exploration of the data.

#### Helpful packages and preliminaries

#Data processing and evaluation
import pandas as pd
import numpy as np

#Graphical display
import matplotlib.pyplot as plt
import seaborn as sns

#Graphing magics
%matplotlib inline

#Constants/globals
cleaned_data_filename = ''

## Assessment and visualization

#load and view data
data = pd.read_csv(cleaned_data_filename)
data.head(10)

#view information about the data
data.info()

#view info about numerical data
data.describe()

#view correlation information
pp = sns.pairplot(data)

#view distribution about numerical columns
vp = sns.violinplot(data)
