==> Asurion-customer-propensity/30-feature-engineering.ipynb <==
# 30-feature-engineering
> Generating new features for the dataset

In this notebook, we generate new features to be used in the dataset.  Keep in mind that depending on the framework you're using, you may want to generate/validate code/functions here to be used as a part of a pipeline in the 40-modeling notebook.

#### Common helpful packages

#Data analysis and processing
import pandas as pd
import numpy as np

#plotting
import matplotlib.pyplot as plt
import seaborn as sns

#Assertions and testing
import great_expectations as ge

# Feature investigation and creation
