==> Asurion-customer-propensity/31-initial-fe.ipynb <==
# Initial Feature Engineering

import pandas as pd
import janitor

%cd ..

### Checking out March Info

march_encoded = pd.read_csv("data/tabular_march_encoded.csv")

march_collated = pd.read_csv("data/transcript_march_collated_encoded.csv")

march_encoded = march_encoded.drop("Unnamed: 0", axis = 1)
march_collated = march_collated.drop("Unnamed: 0", axis = 1)

march_encoded.info()

march_encoded.iloc[:,:15].head()

march_collated.info()

march_collated.head()

### Data Transformation

march_encoded["clndr_dt_cst"] = pd.to_datetime(march_encoded["clndr_dt_cst"], format = "%Y-%m-%d")

march_encoded["start_dt"] = pd.to_datetime(march_encoded["start_dt"], format = "%Y-%m-%d")

march_encoded["claritas_hh_count"] = march_encoded["claritas_hh_count"].astype(int)

march_encoded["zipcode"] = march_encoded["zipcode"].astype('Int64')

columns_to_int = ['seven_day_churn', 'thirty_day_churn','ninety_day_churn', 'lines', 'account_tenure',
                  'num_decline', 'claritas_hh_count','zipcode']

def float_to_int(month, columns):
    month_encoded = pd.read_csv("data/tabular_" + month + "_encoded.csv")
    for i in range(len(columns)):
        month_encoded[columns[i]] = month_encoded[columns[i]].astype('Int64')
    month_encoded["month"] = month
    month_encoded = month_encoded.drop("Unnamed: 0", axis = 1)
    return month_encoded

march_encoded = float_to_int("march",columns_to_int)
april_encoded = float_to_int("april",columns_to_int)
may_encoded = float_to_int("may",columns_to_int)
june_encoded = float_to_int("june",columns_to_int)
july_encoded = float_to_int("july",columns_to_int)

july_encoded.head()

### Adding zipcode information to data sets

zip_code = pd.read_table("data/US.txt")

zip_code = zip_code.drop(["Country","admin_name1","admin_code1","admin_name2","admin_code2"], axis = 1)

zip_code["Zipcode"] = zip_code["Zipcode"].astype("Int64")

def zip_code_adder(month_encoded):
    month_encoded = zip_code.merge(month_encoded, how = "right", left_on = "Zipcode", right_on = "zipcode")
    month_encoded = month_encoded.drop("Zipcode", axis = 1)
    month_encoded = month_encoded.clean_names()
    return month_encoded

march_encoded = zip_code_adder(march_encoded)
april_encoded = zip_code_adder(april_encoded)
may_encoded = zip_code_adder(may_encoded)
june_encoded = zip_code_adder(june_encoded)
july_encoded = zip_code_adder(july_encoded)

### Saving all encoded datasets to fe_data folder

march_encoded.to_csv("fe_data/tabular_march_encoded.csv", index = False)
april_encoded.to_csv("fe_data/tabular_april_encoded.csv", index = False)
may_encoded.to_csv("fe_data/tabular_may_encoded.csv", index = False)
june_encoded.to_csv("fe_data/tabular_june_encoded.csv", index = False)
july_encoded.to_csv("fe_data/tabular_july_encoded.csv", index = False)

### Unique ID comparison

#### Concatenating all tabular datasets

### Concatenating all datasets
total_encoded_tabular = pd.concat([march_encoded, april_encoded, may_encoded, june_encoded, july_encoded], axis = 0)
total_encoded_tabular.info()

total_encoded_tabular.to_csv("fe_data/total_encoded_tabular.csv", index = False)

#### Feature enineering for all transcript datasets

march_collated = pd.read_csv("data/transcript_march_collated_encoded.csv")
april_collated = pd.read_csv("data/transcript_april_collated_encoded.csv")
may_collated = pd.read_csv("data/transcript_may_collated_encoded.csv")
june_collated = pd.read_csv("data/transcript_june_collated_encoded.csv")
july_collated = pd.read_csv("data/transcript_july_collated_encoded.csv")

def transform_transcript(data, month):
    data = data.drop("Unnamed: 0", axis = 1)
    data["month"] = month
    data["sales_offer_date"] = pd.to_datetime(data["sales_offer_date"], format = "%Y-%m-%d")
    data = data.clean_names()
    return data

march_collated = transform_transcript(march_collated, "march")
april_collated = transform_transcript(april_collated, "april")
may_collated = transform_transcript(may_collated, "may")
june_collated = transform_transcript(june_collated, "june")
july_collated = transform_transcript(july_collated, "july")

total_collated_transcripts = pd.concat([march_collated, april_collated, may_collated, june_collated, july_collated], axis = 0)

total_collated_transcripts

total_collated_transcripts.to_csv("fe_data/total_collated_transcripts.csv", index = False)

march_collated.to_csv("fe_data/transcript_march_collated_encoded.csv")
april_collated.to_csv("fe_data/transcript_april_collated_encoded.csv")
may_collated.to_csv("fe_data/transcript_may_collated_encoded.csv")
june_collated.to_csv("fe_data/transcript_june_collated_encoded.csv")
july_collated.to_csv("fe_data/transcript_july_collated_encoded.csv")

### New ID proportions

id_tabular = total_encoded_tabular["id"].unique()
id_transcript = total_collated_transcripts["id"].unique()
id_intersection = list(set(id_tabular).intersection(id_transcript))

encoded_retained = len(id_intersection)/len(id_tabular)
print("%.2f percent of the tabular encoded data were retained."%(encoded_retained*100))

collated_retained = len(id_intersection)/len(id_transcript)
print("%.2f percent of the transcript collated data were retained."%(collated_retained*100))
