==> Asurion-customer-propensity/32-income-fe.ipynb <==
import pandas as pd
import seaborn as sns

%cd ..

march_encoded = pd.read_csv("fe_data/tabular_march_encoded.csv")

march_encoded.info()

march_encoded['hh_income'].value_counts()

march_encoded['hh_income_code'] = march_encoded['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

march_encoded['hh_income_code'].value_counts()

april_encoded = pd.read_csv("fe_data/tabular_april_encoded.csv")

april_encoded['hh_income_code'] = april_encoded['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

april_encoded['hh_income_code'].value_counts()

june_encoded = pd.read_csv("fe_data/tabular_june_encoded.csv")

june_encoded['hh_income_code'] = june_encoded['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

june_encoded['hh_income_code'].value_counts()

may_encoded = pd.read_csv("fe_data/tabular_may_encoded.csv")

may_encoded['hh_income_code'] = may_encoded['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

may_encoded['hh_income_code'].value_counts()

july_encoded = pd.read_csv("fe_data/tabular_july_encoded.csv")

july_encoded['hh_income_code'] = july_encoded['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

july_encoded['hh_income_code'].value_counts()

march_encoded.to_csv("fe_data/tabular_march_encoded.csv", index = False)
april_encoded.to_csv("fe_data/tabular_april_encoded.csv", index = False)
may_encoded.to_csv("fe_data/tabular_may_encoded.csv", index = False)
june_encoded.to_csv("fe_data/tabular_june_encoded.csv", index = False)
july_encoded.to_csv("fe_data/tabular_july_encoded.csv", index = False)

total_encoded_tabular = pd.concat([march_encoded, april_encoded, may_encoded, june_encoded, july_encoded], axis = 0)

total_encoded_tabular.to_csv("fe_data/total_encoded_tabular.csv", index = False)
