==> Asurion-customer-propensity/33-date-fe.ipynb <==
# Feature Engineering on Dates

import pandas as pd

%cd ..

### Date Transformation for all encoded data

#### There was also transformation for some variables that needed their data types refined

def date_transform(data):
    df = pd.read_csv(data)
    df["clndr_dt_cst"] = pd.to_datetime(df["clndr_dt_cst"], format = "%Y-%m-%d")
    df["start_dt"] = pd.to_datetime(df["start_dt"], format = "%Y-%m-%d")
    df["zipcode"] = df["zipcode"].astype('Int64')
    df = df.drop(["lat_long_source","accepted_flg_1"], axis = 1)
    df["day_of_week"] = df['clndr_dt_cst'].dt.day_name()
    df["time_of_day"] = df['start_dt'].dt.hour
    df.to_csv(data, index = False)

%cd data

%cd ..

%cd fe_data

%ls

import os

data_list = os.listdir()

new_data_list = []
for i in range(len(data_list)):
    if "collated" not in data_list[i]:
        new_data_list.append(data_list[i])

for data in new_data_list:
    date_transform(data)

%cd ..

march_encoded = pd.read_csv("fe_data/tabular_march_encoded.csv")
march_encoded

march_encoded.info()
