==> Asurion-customer-propensity/35-fe-column-rename.ipynb <==
import pandas as pd

%cd ..

df_encoded = pd.read_csv("fe_data/total_encoded_tabular.csv")

df_encoded.info()

df_encoded = df_encoded.rename(columns = {"zipcode_total_sales":"zipcode_total_offers",
                                          "zipcode_neg_sales":"zipcode_rejected_offers",
                                          "zipcode_pos_sales":"zipcode_accepted_offers",
                                          "zipcode_avg_sales":"zipcode_accepted_prop"})
df_encoded.info()

df_encoded.to_csv("fe_data/total_encoded_tabular.csv")

def month_storage(df, month):
    month_data = df[df["month"] == month]
    month_data.to_csv("fe_data/tabular_" + month + "_encoded.csv")

month_storage(df_encoded, "march")
month_storage(df_encoded, "april")
month_storage(df_encoded, "may")
month_storage(df_encoded, "june")
month_storage(df_encoded, "july")
