==> Asurion-customer-propensity/36-filtering.ipynb <==
# 36-filtering

Filter out only call transcripts of customers who accepted the upsale offer

%cd ..

import pandas as pd

df = pd.read_csv('fe_data/total_merged_asurion.csv')
df.head()

df.columns

df = df.rename(columns={'accepted_flg_x':'accepted_flg', 'month_x':'month'})

df.info()

aqs = df[(df['quality_sale'] == 1) & (df['accepted_flg'] == 1)]
aqs.drop_duplicates().to_csv('fe_data/accepted_upsale_merged_asurion.csv', index=0)

aqs.columns

atrans = aqs[['id', 'sales_offer_date', 'accepted_flg',
       'encrypted_collated_transcription', 'month']]
atrans.drop_duplicates().to_csv('fe_data/accepted_upsale_collated_transcripts.csv', index=0)

atrans.columns

atabular = aqs[['id', 'quality_sale', 'accepted_flg', 'seven_day_churn',
       'thirty_day_churn', 'ninety_day_churn', 'clndr_dt_cst', 'start_dt',
       'lines', 'previous_tech_support_call', 'account_tenure',
       'past_yr_claim_count', 'num_decline', 'exp', 'product_type', 'pmake',
       'claritas_hh_count', 'technodoption_class', 'hh_income', 'age_class',
       'hh_education', 'hh_composition', 'hh_tenure', 'urbanicity',
       'median_income', 'total_population', 'perc_female', 'perc_under18',
       'perc_over65', 'perc_white', 'perc_black', 'population_density',
       'zipcode', 'county', 'state', 'state_code', 'latitude', 'longitude',
       'month', 'hh_income_code', 'day_of_week', 'time_of_day',
       'zipcode_accepted_prop', 'zipcode_rejected_offers',
       'zipcode_accepted_offers', 'zipcode_total_offers']]
atabular.drop_duplicates().to_csv('fe_data/accepted_upsale_encoded_tabular.csv', index=0)

atabular.columns
