==> Asurion-customer-propensity/37-standardization.ipynb <==
# 37-standardization

import pandas as pd

%cd ..

total_encoded_tabular = pd.read_csv('fe_data/total_encoded_tabular.csv', index_col='id')
total_encoded_tabular = total_encoded_tabular.drop(columns='Unnamed: 0')
total_encoded_tabular.head()

total_encoded_tabular.info()

total_encoded_tabular.shape

total_encoded_tabular.iloc[0:5, :10]

# len(total_encoded_tabular[total_encoded_tabular['quality_sale']==1])/len(total_encoded_tabular)

df = total_encoded_tabular[total_encoded_tabular['quality_sale']==1]
print(df['ninety_day_churn'].value_counts(normalize=True))
print(df['thirty_day_churn'].value_counts(normalize=True))
print(df['seven_day_churn'].value_counts(normalize=True))

total_encoded_tabular.iloc[0:5, 10:20]

total_encoded_tabular.iloc[0:5, 20:30]

total_encoded_tabular.iloc[0:5, 30:]

total_encoded_tabular.columns

flgs = ['quality_sale', 'accepted_flg', 'seven_day_churn',
       'thirty_day_churn', 'ninety_day_churn', 'exp']
dates = ['clndr_dt_cst', 'start_dt']
num_list = ['lines', 'previous_tech_support_call', 'account_tenure', 'past_yr_claim_count', 'num_decline', 
            'claritas_hh_count', 'median_income', 'total_population', 'perc_female', 'perc_under18',
            'perc_over65', 'perc_white', 'perc_black', 'population_density', 'latitude', 'longitude',
            'zipcode_accepted_prop', 'zipcode_rejected_offers', 'zipcode_accepted_offers', 'zipcode_total_offers']
cat_list = ['product_type', 'pmake', 'technodoption_class', 'hh_income', 'age_class', 'hh_education', 
            'hh_composition', 'hh_tenure', 'urbanicity', 'month', 'day_of_week', 'time_of_day']
others = ['zipcode', 'county', 'state', 'state_code', 'hh_income_code']

for i in num_list:
    print(i+' missing rate:',total_encoded_tabular[i].isnull().sum()/len(total_encoded_tabular[i]))

for i in num_list:
    total_encoded_tabular[i] = total_encoded_tabular[i].fillna(0)

total_encoded_tabular.shape[1] == len(flgs) + len(dates) + len(num_list) + len(cat_list) + len(others)

### Categorical Variables

for col in cat_list:
    print(col+':', total_encoded_tabular[col].nunique(), total_encoded_tabular[col].unique(), '\n')

# label encoding
# for col in cat_list:
#     total_encoded_tabular[col] = total_encoded_tabular[col].astype('category')
#     total_encoded_tabular[col+'_cat'] = total_encoded_tabular[col].cat.codes
# for col in cat_list:
#     print(col+':', total_encoded_tabular[col+'_cat'].nunique(), total_encoded_tabular[col+'_cat'].unique(), '\n')

# one hot encoding
total_encoded_tabular = pd.get_dummies(total_encoded_tabular, columns=cat_list, prefix=cat_list)

def rename_col(col:str):
    col = col.lower() # convert names to all lowercase
#     col = re.sub('[{}]'.format(string.punctuation), " ", col) # remove punctuations
    col = ' '.join(col.split()) # fix whitespaces
    col = col.replace(' ', '_') # replace spaces with underscores
    col = col.replace('-', '_')
    return col
for col in total_encoded_tabular.columns:
    total_encoded_tabular.rename(columns={col:rename_col(col)}, inplace=True)

### Numerical Variables

def mynorm(df, num_list, norm_method):
    # norm_method = True zscore, False min-max
    for col in num_list:
        df[col+'_normalized'] = df[col].transform(lambda x: (x - x.mean()) / x.std()) if norm_method == True else df[col].transform(lambda x: (x - x.min()) / (x.max() - x.min()))
    return df


# normalize numerical columns
total_encoded_tabular = mynorm(total_encoded_tabular, num_list, False)

total_encoded_tabular = total_encoded_tabular.drop(columns = (num_list+dates+others))

check=pd.DataFrame(total_encoded_tabular.isnull().sum(), columns=['count'])
check[check['count']!=0]

# total_encoded_tabular.to_csv('fe_data/total_encoded_standardized.csv')
total_encoded_tabular.to_csv('fe_data/total_encoded_standardized_fillna.csv')

# label encoding
# total_encoded_tabular = total_encoded_tabular.drop(columns = (num_list+dates+others))
# total_encoded_tabular.to_csv('fe_data/total_encoded_standardized_w_label.csv')
# total_encoded_tabular_copy = total_encoded_tabular.drop(columns = (cat_list))
# total_encoded_tabular_copy.to_csv('fe_data/total_encoded_standardized.csv')

total_encoded_tabular.shape

total_encoded_tabular.columns

for col in total_encoded_tabular.columns:
    print(col)
