==> Asurion-customer-propensity/38-fe-train-data.ipynb <==
import pandas as pd
import janitor

%cd ..

### Regular FE

def month_import(month):
    df = pd.read_csv("data/tabular_" + month + "_encoded.csv")
    df = df.drop("Unnamed: 0", axis = 1)
    df["month"] = month
    return df

df_march = month_import("march")
df_april = month_import("april")
df_may = month_import("may")
df_june = month_import("june")

df_train = pd.concat([df_march, df_april, df_may, df_june], axis = 0)
df_train.info()

columns_to_int = ['seven_day_churn', 'thirty_day_churn','ninety_day_churn', 'lines', 'account_tenure',
                  'num_decline', 'claritas_hh_count']

for i in range(len(columns_to_int)):
    df_train[columns_to_int[i]] = df_train[columns_to_int[i]].astype('Int64')

df_train.info()

zip_code = pd.read_table("data/US.txt")

zip_code = zip_code.drop(["Country","admin_name1","admin_code1","admin_name2","admin_code2"], axis = 1)

zip_code["Zipcode"] = zip_code["Zipcode"].astype(float)

df_train = zip_code.merge(df_train, how = "right", left_on = "Zipcode", right_on = "zipcode")
df_train = df_train.drop("Zipcode", axis = 1)
df_train = df_train.clean_names()

df_train

### Income FE

df_train['hh_income_code'] = df_train['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

### Date FE

df_train["clndr_dt_cst"] = pd.to_datetime(df_train["clndr_dt_cst"], format = "%Y-%m-%d")
df_train["start_dt"] = pd.to_datetime(df_train["start_dt"], format = "%Y-%m-%d")
df_train = df_train.drop(["lat_long_source","accepted_flg_1"], axis = 1)
df_train["day_of_week"] = df_train['clndr_dt_cst'].dt.day_name()
df_train["time_of_day"] = df_train['start_dt'].dt.hour

### Zipcode Measures

zipcode_df = df_train.groupby(["zipcode","quality_sale"], as_index = False).size()
zipcode_df = zipcode_df.pivot_table(index = "zipcode", values = "size", columns = "quality_sale")
zipcode_df = zipcode_df.fillna(0)

zipcode_df["mean_sales"] = zipcode_df.loc[:,1]/(zipcode_df.loc[:,0] + zipcode_df.loc[:,1])
zipcode_df["zipcode"] = zipcode_df.index
zipcode_df = zipcode_df.reset_index(drop = True)
zipcode_df["total_sales"] = zipcode_df[0] + zipcode_df[1]
zipcode_df = zipcode_df.rename_axis(None, axis=1)

zipcode_df

df_train_zipcode = df_train.merge(zipcode_df, how = "left", on = "zipcode")

df_train_zipcode = df_train_zipcode.rename(columns = {0:"zipcode_neg_sales",1:"zipcode_pos_sales","total_sales":"zipcode_total_sales","mean_sales":"zipcode_avg_sales"})

df_train_zipcode = df_train_zipcode[['id','quality_sale', 'accepted_flg', 'seven_day_churn', 'thirty_day_churn',
                    'ninety_day_churn', 'clndr_dt_cst', 'start_dt', 'lines',
                    'previous_tech_support_call', 'account_tenure', 'past_yr_claim_count',
                    'num_decline', 'exp', 'product_type', 'pmake', 'claritas_hh_count',
                    'technodoption_class', 'hh_income', 'age_class', 'hh_education',
                    'hh_composition', 'hh_tenure', 'urbanicity', 'median_income',
                    'total_population', 'perc_female', 'perc_under18', 'perc_over65',
                    'perc_white', 'perc_black', 'population_density', 'zipcode', 'county', 
                    'state', 'state_code', 'latitude', 'longitude',
                    'month', 'hh_income_code', 'day_of_week', 'time_of_day',
                    'zipcode_avg_sales',  'zipcode_neg_sales', 'zipcode_pos_sales', 'zipcode_total_sales']]

df_train_zipcode = df_train_zipcode.rename(columns = {"zipcode_total_sales":"zipcode_total_offers",
                                          "zipcode_neg_sales":"zipcode_rejected_offers",
                                          "zipcode_pos_sales":"zipcode_accepted_offers",
                                          "zipcode_avg_sales":"zipcode_accepted_prop"})

df_train_zipcode

df_train.to_csv("fe_data_v2/train_encoded.csv")

df_train_zipcode.to_csv("fe_data_v2/train_encoded_zipcode.csv", index = False)

df_train_zipcode.info()
