==> Asurion-customer-propensity/39-fe-test-data.ipynb <==
import pandas as pd
import janitor

%cd ..

### Regular FE

df_test = pd.read_csv("data/tabular_july_encoded.csv").clean_names()
df_test.info()

columns_to_int = ['seven_day_churn', 'thirty_day_churn','ninety_day_churn', 'lines', 'account_tenure',
                  'num_decline', 'claritas_hh_count']

for i in range(len(columns_to_int)):
    df_test[columns_to_int[i]] = df_test[columns_to_int[i]].astype('Int64')

df_test.info()

### Income FE

df_test['hh_income_code'] = df_test['hh_income'].replace(['1 Wealthy', '2 Upscale', '3 Upper Midscale', '4 Midscale', '5 Lower Midscale', '6 Downscale', '7 Low Income', 'Unknown'], [1, 2, 3, 4, 5, 6, 7, 8])

### Date FE

df_test["clndr_dt_cst"] = pd.to_datetime(df_test["clndr_dt_cst"], format = "%Y-%m-%d")
df_test["start_dt"] = pd.to_datetime(df_test["start_dt"], format = "%Y-%m-%d")
df_test = df_test.drop("accepted_flg_1", axis = 1)
df_test["day_of_week"] = df_test['clndr_dt_cst'].dt.day_name()
df_test["time_of_day"] = df_test['start_dt'].dt.hour

df_test = df_test.drop("unnamed_0", axis = 1)

df_test["month"] = "july"

df_test.info()

df_train_zip = pd.read_csv("fe_data_v2/train_encoded_zipcode.csv")

df_train_zip.info()

zipcode_info = df_train_zip[[ "zipcode","county", "state","state_code","latitude","longitude","zipcode_accepted_prop","zipcode_rejected_offers","zipcode_accepted_offers","zipcode_total_offers"]]

zipcode_info

zipcode_info = zipcode_info.drop_duplicates()

zipcode_info.shape

df_test_zip = df_test.merge(zipcode_info, how = "left", on = "zipcode")

df_test_zip

df_test_zip = df_test_zip[['id','quality_sale', 'accepted_flg', 'seven_day_churn', 'thirty_day_churn',
                    'ninety_day_churn', 'clndr_dt_cst', 'start_dt', 'lines',
                    'previous_tech_support_call', 'account_tenure', 'past_yr_claim_count',
                    'num_decline', 'exp', 'product_type', 'pmake', 'claritas_hh_count',
                    'technodoption_class', 'hh_income', 'age_class', 'hh_education',
                    'hh_composition', 'hh_tenure', 'urbanicity', 'median_income',
                    'total_population', 'perc_female', 'perc_under18', 'perc_over65',
                    'perc_white', 'perc_black', 'population_density', 'zipcode', 'county', 
                    'state', 'state_code', 'latitude', 'longitude',
                    'month', 'hh_income_code', 'day_of_week', 'time_of_day',
                    'zipcode_accepted_prop',  'zipcode_rejected_offers', 'zipcode_accepted_offers', 'zipcode_total_offers']]

df_test_zip.info()

df_test.to_csv("fe_data_v2/test_encoded.csv", index = False)

df_test_zip.to_csv("fe_data_v2/test_encoded_zipcode.csv", index = False)
