==> Asurion-customer-propensity/48-modeling-catboost-30day-sentiment.ipynb <==
Catboost - 30 day churn - sentiment

!pip install tensorflow

from tensorflow.python.client import device_lib

def get_available_devices():
    local_device_protos = device_lib.list_local_devices()
    return [x.name for x in local_device_protos]

print(get_available_devices())

!pip install catboost

import pandas as pd
import numpy as np
# import seaborn
import matplotlib.pyplot as plt
%matplotlib inline

import pickle
import joblib
import time

from catboost import CatBoostClassifier

import xgboost as xgb
from sklearn.model_selection import train_test_split, GridSearchCV
from xgboost import XGBClassifier, plot_importance
from sklearn import config_context
from sklearn.metrics import classification_report, confusion_matrix, ConfusionMatrixDisplay
from sklearn.metrics import accuracy_score, mean_squared_error
from sklearn.preprocessing import MinMaxScaler
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.model_selection import train_test_split 
from sklearn.model_selection import cross_val_predict 
from sklearn.model_selection import cross_val_score 
from sklearn.feature_selection import RFECV
from sklearn.metrics import precision_score, f1_score, recall_score, confusion_matrix, accuracy_score, roc_auc_score

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 500)
pd.set_option('display.float_format', '{:.2f}'.format)

!pip install ipywidgets
!pip install catboost
import ipywidgets
import catboost as cb
from catboost import CatBoostClassifier
from sklearn.metrics import auc, roc_auc_score, roc_curve

import warnings
warnings.filterwarnings('ignore')

%cd ..

test = pd.read_csv('fe_data_v2/30_day_churn_data/test_30d_encoded.csv')

sen_test = pd.read_csv('fe_data_v2/sentiment_data/sentiment_test.csv')

len(sen_test)

len(test)

test.columns

dcols = ['quality_sale', 'clndr_dt_cst', 'start_dt', 'zipcode', 'county', 'state', 
         'state_code', 'latitude', 'longitude', 'month','hh_income_code']

test.drop(columns = dcols, inplace = True)
test = test[(test['zipcode_accepted_prop'].isnull()==False)& (test['zipcode_rejected_offers'].isnull()==False)& 
            (test['zipcode_accepted_offers'].isnull()==False)&(test['zipcode_total_offers'].isnull()==False)]

len(sen_test)

test = test.merge(sen_test, on='id', how='inner')

len(test)

test.info()

train = pd.read_csv('fe_data_v2/30_day_churn_data/train_30d_encoded.csv')

len(train)

sen_train = pd.read_csv('fe_data_v2/sentiment_data/sentiment_train.csv')

len(sen_train["id"].unique())

train = train.merge(sen_train, on='id', how='inner')

len(train)

train.info()

train.head()

len(train)

train.drop(columns = dcols, inplace = True)
train = train[(train['zipcode_accepted_prop'].isnull()==False)& (train['zipcode_rejected_offers'].isnull()==False)& 
            (train['zipcode_accepted_offers'].isnull()==False)&(train['zipcode_total_offers'].isnull()==False)]

len(train)

train.columns

test.columns

def transform_xy(dat):
    X_ori = dat[[
#         'emplid','cvp_call_id',
                 'lines', 'previous_tech_support_call','account_tenure', 'past_yr_claim_count', 
                'num_decline', 'exp', 
#                  'churn',
                 'product_type', 'pmake', 
                 'claritas_hh_count',
                 'technodoption_class', 'hh_income', 'age_class', 'hh_education',
                 'hh_composition', 'hh_tenure', 'urbanicity','median_income', 'total_population', 
                 'perc_female', 'perc_under18',
                'perc_over65', 'perc_white', 'perc_black', 'population_density', 'day_of_week', 'time_of_day',
       'zipcode_accepted_prop', 'zipcode_rejected_offers',
       'zipcode_accepted_offers', 'zipcode_total_offers', 'sentiment']]
    y = dat[['thirty_day_churn']]
   
    X_ori.loc[X_ori.lines.isnull(), 'lines'] = 0
    X_ori.exp = X_ori.exp.apply(int)
    X_ori.num_decline = X_ori.num_decline.apply(float)
    X_ori.loc[X_ori.account_tenure.isnull(), 'account_tenure'] = 0
    X_ori.loc[X_ori.claritas_hh_count.isnull(), 'claritas_hh_count'] = 0
    X_ori.loc[X_ori.exp.isnull(), 'exp'] = 'Unknown'
    X_ori.loc[X_ori.product_type.isnull(), 'product_type'] = 'Unknown'
    X_ori.loc[X_ori.pmake.isnull(), 'pmake'] = 'Unknown'
    X_ori.loc[X_ori.technodoption_class.isnull(), 'technodoption_class'] = 'Unknown'
    X_ori.loc[X_ori.hh_income.isnull(), 'hh_income'] = 'Unknown'
    X_ori.loc[X_ori.age_class.isnull(), 'age_class'] = 'Unknown'
    X_ori.loc[X_ori.hh_education.isnull(), 'hh_education'] = 'Unknown'
    X_ori.loc[X_ori.hh_composition.isnull(), 'hh_composition'] = 'Unknown'
    X_ori.loc[X_ori.hh_tenure.isnull(), 'hh_tenure'] = 'Unknown'
    X_ori.loc[X_ori.urbanicity.isnull(), 'urbanicity'] = 'Unknown'
    X_ori.loc[X_ori.day_of_week.isnull(), 'day_of_week'] = 'Unknown'
    X_ori.loc[X_ori.time_of_day.isnull(), 'time_of_day'] = 'Unknown'
    X_ori.loc[X_ori.zipcode_accepted_prop.isnull(), 'zipcode_accepted_prop'] = 0
    X_ori.loc[X_ori.zipcode_rejected_offers.isnull(), 'zipcode_rejected_offers'] = 0
    X_ori.loc[X_ori.zipcode_accepted_offers.isnull(), 'zipcode_accepted_offers'] = 0
    X_ori.loc[X_ori.zipcode_total_offers.isnull(), 'zipcode_total_offers'] = 0
    X_ori.loc[X_ori.sentiment.isnull(), 'sentiment'] = "Unknown"

    return X_ori,y

X,y = transform_xy(train)
x_hold_out, y_hold_out = transform_xy(test)

X['lines'] = X['lines'].astype(np.int) 
X['previous_tech_support_call'] = X['previous_tech_support_call'].astype('category')
X['num_decline'] = X['num_decline'].astype(np.int) 
X['past_yr_claim_count'] = X['past_yr_claim_count'].astype(np.int) 
X['exp'] = X['exp'].astype('category') 
X['account_tenure'] = X['account_tenure'].astype(np.int) 
X['pmake'] = X['pmake'].astype('category')
X['technodoption_class'] = X['technodoption_class'].astype('category')
X['claritas_hh_count'] = X['claritas_hh_count'].astype('category')
X['hh_income'] = X['hh_income'].astype('category')
X['age_class'] = X['age_class'].astype('category')
X['hh_education'] = X['hh_education'].astype('category')
X['hh_composition'] = X['hh_composition'].astype('category')
X['hh_tenure'] = X['hh_tenure'].astype('category')
X['urbanicity'] = X['urbanicity'].astype('category')
X['product_type'] = X['product_type'].astype('category')
X['day_of_week'] = X['day_of_week'].astype('category')
X['time_of_day'] = X['time_of_day'].astype('category')
X['zipcode_accepted_prop'] = (X['zipcode_accepted_prop']*100).astype(np.int)
X['zipcode_rejected_offers'] = X['zipcode_rejected_offers'].astype(np.int)
X['zipcode_accepted_offers'] = X['zipcode_accepted_offers'].astype(np.int)
X['zipcode_total_offers'] = X['zipcode_total_offers'].astype(np.int)
X['sentiment'] = X['sentiment'].astype('category')


X['median_income'] = X['median_income'].astype(np.int)
X['total_population'] = X['total_population'].astype(np.int)
X['perc_female'] = X['perc_female'].astype(np.int)
X['perc_under18'] = X['perc_under18'].astype(np.int)
X['perc_over65'] = X['perc_over65'].astype(np.int)
X['perc_white'] = X['perc_white'].astype(np.int)
X['perc_black'] = X['perc_black'].astype(np.int)
X['population_density'] = X['population_density'].astype(np.int)

x_hold_out['lines'] = x_hold_out['lines'].astype(np.int) 
x_hold_out['previous_tech_support_call'] = x_hold_out['previous_tech_support_call'].astype('category')
x_hold_out['num_decline'] = x_hold_out['num_decline'].astype(np.int) 
x_hold_out['past_yr_claim_count'] = x_hold_out['past_yr_claim_count'].astype(np.int) 
x_hold_out['exp'] = x_hold_out['exp'].astype('category') 
x_hold_out['account_tenure'] = x_hold_out['account_tenure'].astype(np.int) 
x_hold_out['pmake'] = x_hold_out['pmake'].astype('category')
x_hold_out['technodoption_class'] = x_hold_out['technodoption_class'].astype('category')
x_hold_out['claritas_hh_count'] = x_hold_out['claritas_hh_count'].astype('category')
x_hold_out['hh_income'] = x_hold_out['hh_income'].astype('category')
x_hold_out['age_class'] = x_hold_out['age_class'].astype('category')
x_hold_out['hh_education'] = x_hold_out['hh_education'].astype('category')
x_hold_out['hh_composition'] = x_hold_out['hh_composition'].astype('category')
x_hold_out['hh_tenure'] = x_hold_out['hh_tenure'].astype('category')
x_hold_out['urbanicity'] = x_hold_out['urbanicity'].astype('category')
x_hold_out['product_type'] = x_hold_out['product_type'].astype('category')
x_hold_out['day_of_week'] = x_hold_out['day_of_week'].astype('category')
x_hold_out['time_of_day'] = x_hold_out['time_of_day'].astype('category')
x_hold_out['zipcode_accepted_prop'] = (x_hold_out['zipcode_accepted_prop']*100).astype(np.int)
x_hold_out['zipcode_rejected_offers'] = x_hold_out['zipcode_rejected_offers'].astype(np.int)
x_hold_out['zipcode_accepted_offers'] = x_hold_out['zipcode_accepted_offers'].astype(np.int)
x_hold_out['zipcode_total_offers'] = x_hold_out['zipcode_total_offers'].astype(np.int)
x_hold_out['sentiment'] = x_hold_out['sentiment'].astype('category')


x_hold_out['median_income'] = x_hold_out['median_income'].astype(np.int)
x_hold_out['total_population'] = x_hold_out['total_population'].astype(np.int)
x_hold_out['perc_female'] = x_hold_out['perc_female'].astype(np.int)
x_hold_out['perc_under18'] = x_hold_out['perc_under18'].astype(np.int)
x_hold_out['perc_over65'] = x_hold_out['perc_over65'].astype(np.int)
x_hold_out['perc_white'] = x_hold_out['perc_white'].astype(np.int)
x_hold_out['perc_black'] = x_hold_out['perc_black'].astype(np.int)
x_hold_out['population_density'] = x_hold_out['population_density'].astype(np.int)

len(x_hold_out)

len(X)

len(y)

type(train)

type(test)

len(test)

len(x_hold_out)

len(y_hold_out)

Xf_train, Xf_test, y_train, y_test = train_test_split(X,y, test_size = 0.3, stratify=y['thirty_day_churn'])

Xf_train.columns

Xf_test.columns

spw = (y_train['thirty_day_churn'] == 0.00).sum()/(y_train['thirty_day_churn'] == 1.00).sum()
spw

train_census_dataset = cb.Pool(Xf_train,y_train, 
                        cat_features=np.where(Xf_train.dtypes != np.int)[0])                                                      
test_census_dataset = cb.Pool(Xf_test, y_test,           
                       cat_features=np.where(Xf_test.dtypes != np.int)[0])

holdout_census_dataset = cb.Pool(x_hold_out, y_hold_out,           
                       cat_features=np.where(x_hold_out.dtypes != np.int)[0])

from catboost import CatBoostClassifier

#initiate model
cat_model_census = CatBoostClassifier(
    cat_features = np.where(Xf_train.dtypes != np.int)[0],
    scale_pos_weight=spw,
    max_depth = 5,
    learning_rate = 0.05
)
#train
cat_model_census.fit(train_census_dataset, 
       eval_set=test_census_dataset,
      plot=True, silent= True)

pred_census = cat_model_census.predict_proba(holdout_census_dataset)
threshold=0.5
predicted_cat_census = (pred_census[:,1] >= threshold).astype('int')
tn, fp, fn, tp = confusion_matrix(y_hold_out, predicted_cat_census).ravel()
print("catboost_holdoutset",'\n',
    "recall:",recall_score(y_hold_out, predicted_cat_census),'\n',
     'precision',precision_score(y_hold_out, predicted_cat_census),'\n',
     confusion_matrix(y_hold_out, predicted_cat_census),'\n',
     'f1 score:',f1_score(y_hold_out, predicted_cat_census),'\n',
     'tn:{}, fp:{}, fn:{}, tp:{}'.format(tn, fp, fn, tp),'\n')

pred_census = cat_model_census.predict_proba(test_census_dataset)
threshold=0.5
predicted_cat_census = (pred_census[:,1] >= threshold).astype('int')
tn, fp, fn, tp = confusion_matrix(y_test, predicted_cat_census).ravel()
print("catboost_testset",'\n',
    "recall:",recall_score(y_test, predicted_cat_census),'\n',
     'precision',precision_score(y_test, predicted_cat_census),'\n',
     confusion_matrix(y_test, predicted_cat_census),'\n',
     'f1 score:',f1_score(y_test, predicted_cat_census),'\n',
     'tn:{}, fp:{}, fn:{}, tp:{}'.format(tn, fp, fn, tp),'\n')

vip = cat_model_census.feature_importances_

#create df with vip info
feature_importances_info = pd.DataFrame({'feat_names':Xf_train.columns, 'vip': vip})

#get sign and magnitude information
feature_importances_info = feature_importances_info.assign(feature_importances_mag = abs(feature_importances_info['vip']),
                             feature_importances_sign = np.sign(feature_importances_info['vip']))
feature_importances_info = feature_importances_info.set_index('feat_names').sort_values(by='feature_importances_mag', ascending=False)
feature_importances_info.head(15)

!pip install seaborn

import seaborn as sns
sns.barplot(y=feature_importances_info.index[:10], x='feature_importances_mag', hue='feature_importances_sign', 
            data=feature_importances_info[:10], orient='h', dodge=False);

plt.figure(figsize = (16,8))
cor = train.corr()
sns.heatmap(cor, annot = True, cmap = "viridis")
plt.show()
