==> Asurion-customer-propensity/50-modeling-h2o.ipynb <==
# 50-modeling-h2o
> Training models using h2o

In this notebook series, we train models using the h2o modeling package.

#### Common helpful packages

#Data analysis and processing
import pandas as pd
import numpy as np

#machine learning
import h2o

#plotting
import matplotlib.pyplot as plt
import seaborn as sns

#Constants/globals
cleaned_data_filename = ''

#for machine learning
training_ratio = 0.7
max_models = None
repro_seed = 1234
max_runtime_mins = 5

# Initialize h2o and configure data

#init h2o
h2o.init()

#read the data from file
df = pd.read_csv(cleaned_data_filename)

#convert to h2o format
h2o_df = h2o.H2OFrame(df)

# Generate models using h2o

#split data into desired portions and identify variables of interest
data_train, data_val = h2o_df.split_frame(ratios=(training_ratio), seed=repro_seed)

predictors = ['col1', 'col2']
response = 'coly'

#set columns as factors if necessary
data['col'] = data['col'].as_factor()

#identify model and train
mls = H2OAutoML(max_models=max_models, max_runtime_secs = 60* max_runtime_mins, seed=repro_seed)
mls.train(x=predictors, y=response, training_frame=data_train)

#View leaderboard
mls.leaderboard

### h2o performance evaluation

#See full performance of model on original training data
mls.leader.model_performance(train_data)

#Predict outputs using leader (usage of the leader done automatically)
res = mls.predict(test_data)

#See full performance of model
mls.leader.model_performance(test_data)

#Save model
