==> Asurion-customer-propensity/70-modeling-fastai.ipynb <==
# 70-modeling-fastai
> Training models

In this notebook series, we train models using fastai.

#### Common helpful packages

#Data analysis and processing
import pandas as pd
import numpy as np

#plotting
import matplotlib.pyplot as plt
import seaborn as sns

#Constants/globals
cleaned_data_filename = ''

## Modeling via deep learning

#### Helpful packages and preliminaries

#fast dev dl packages
import fastai as fa
from fastai.text import * as fat
from fastai.callbacks import * as fac

# Constants/globals
bs = 128
path = %pwd
model_save_dir = Path(path)/'models'
db_save_dir = Path(path)/'databunches'

#Hot fix for fast.ai v1.0.57 to fix issue about fast.ai not creating databunch file
import os
os.mkdir(databunch_dir)

#Load constants

#Read in as databunch

#Save databunch

# Model type and parameters

# Train model

# Test model

# Save model

### dl performance evaluation
