==> Asurion-customer-propensity/90-inference.ipynb <==
# 90-inference
> Using built models for inference

In this notebook series, we use models for inference.  This notebook can be used as a demo or for collaborators to create new predictions.  Here, we show an example using huggingface.

#### Common helpful packages

#Data analysis and processing
import pandas as pd
import numpy as np

#ml and dl
from transformers import pipeline

#plotting
import matplotlib.pyplot as plt
import seaborn as sns

# Load data

# Predict using model
Here, we'll just directly input some values.

qa_model = pipeline('question-answering')
answers = qa_model(context='My name is John and I am 13 years old', question='What is my name?')
answers
